requirements.txt
aiokafka==0.13.0
aioredis==2.0.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
async-timeout==5.0.1
asyncpg==0.31.0
bcrypt==5.0.0
bidict==0.23.1
black==26.1.0
blinker==1.9.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
ConfigArgParse==1.7.1
coverage==7.13.4
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.131.0
flake8==7.3.0
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.8
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.43.3
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.9.2
pluggy==1.6.0
prometheus_client==0.24.1
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.1
python-engineio==4.13.1
python-jose==3.5.0
python-socketio==5.16.1
pytokens==0.4.1
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.2.0
requests==2.32.5
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
wsproto==1.3.2
zope.event==6.1
zope.interface==8.2
